requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
langdetect==1.0.9
multidict==6.6.2
playwright==1.53.0
propcache==0.3.2
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
xml-python==0.4.3
yarl==1.20.1
